requirements.txt
torch>=2.0.0
transformers>=4.30.0
torch-geometric>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
biopython>=1.81
plotly>=5.18.0
networkx>=3.1
kaleido>=0.2.1